More work on the mining AI

+++ src/artillects/entity/ai/EntityAIMining.java
@@ -6,6 +6,7 @@ import net.minecraft.block.Block;
 import net.minecraft.entity.ai.EntityAIBase;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
+import artillects.entity.BlockScanner;
 import artillects.entity.EntityWorker;
 
 public class EntityAIMining extends EntityAIBase
@@ -21,57 +22,34 @@ public class EntityAIMining extends EntityAIBase
 	private int breakingTime;
 	private int maxBreakTime = -1;
 
+	private BlockScanner scannedBlocks;
+
 	public EntityAIMining(EntityWorker entity, double par2)
 	{
 		this.entity = entity;
 		this.world = entity.worldObj;
 		this.moveSpeed = par2;
-		// this.setMutexBits(1);
 	}
 
 	/** Returns whether the EntityAIBase should begin execution. */
 	@Override
 	public boolean shouldExecute()
 	{
-		/** Check all blocks under this to find ores. */
-		int x = (int) this.entity.posX;
-		int y = (int) this.entity.posY;
-		int z = (int) this.entity.posZ;
-
-		for (; y > 0; y--)
-		{
-			int blockID = this.world.getBlockId(x, y, z);
-
-			// if (blockID == Block.oreIron.blockID)
-			{
-				return !this.entity.isInventoryFull();
-			}
-		}
-
-		return false;
+		return !this.entity.isInventoryFull();
 	}
 
 	/** Returns whether an in-progress EntityAIBase should continue executing */
 	@Override
 	public boolean continueExecuting()
 	{
+		return this.scannedBlocks.scannedPositions.size() > 0 && !this.entity.isInventoryFull();
+	}
 
-		/** Check all blocks under this to find ores. */
-		int x = (int) this.entity.posX;
-		int y = (int) this.entity.posY;
-		int z = (int) this.entity.posZ;
-
-		for (; y > 0; y--)
-		{
-			int blockID = this.world.getBlockId(x, y, z);
-
-			// if (blockID == Block.oreIron.blockID)
-			{
-				return !this.entity.isInventoryFull();
-			}
-		}
-
-		return false;
+	@Override
+	public void startExecuting()
+	{
+		this.scannedBlocks = new BlockScanner(this.world, entity.zone, Block.oreIron, Block.oreCoal, Block.oreRedstone);
+		this.scannedBlocks.calculate();
 	}
 
 	/** Resets the task */
@@ -79,44 +57,50 @@ public class EntityAIMining extends EntityAIBase
 	public void resetTask()
 	{
 		this.breakingTime = 0;
+		this.scannedBlocks.calculate();
+		System.out.println("reset"+this.scannedBlocks.scannedPositions.size());
 	}
 
 	/** Updates the task */
 	@Override
 	public void updateTask()
 	{
-		this.breakingTime++;
-		int i = (int) (this.breakingTime / 240f * 10f);
-
-		int breakX = (int) this.entity.posX;
-		int breakY = (int) this.entity.posY - 1;
-		int breakZ = (int) this.entity.posZ;
-
-		int blockID = this.world.getBlockId(breakX, breakY, breakZ);
-
-		if (blockID != 0)
+		if (this.scannedBlocks.scannedPositions.size() > 0)
 		{
+			int breakX = (int) this.scannedBlocks.scannedPositions.get(0).x;
+			int breakY = (int) this.scannedBlocks.scannedPositions.get(0).y;
+			int breakZ = (int) this.scannedBlocks.scannedPositions.get(0).z;
 
-			if (i != this.maxBreakTime)
+			this.entity.getNavigator().tryMoveToXYZ(breakX, breakY, breakZ, this.moveSpeed);
+
+			this.breakingTime++;
+			int i = (int) (this.breakingTime / 240f * 10f);
+
+			int blockID = this.world.getBlockId(breakX, breakY, breakZ);
+
+			if (blockID != 0)
 			{
-				this.world.destroyBlockInWorldPartially(this.entity.entityId, breakX, breakY, breakZ, i);
-			}
 
-			if (this.breakingTime == 240)
-			{
-				List<ItemStack> droppedStacks = Block.blocksList[blockID].getBlockDropped(world, breakX, breakY, breakZ, this.world.getBlockMetadata(breakX, breakY, breakZ), 0);
-
-				for (ItemStack stack : droppedStacks)
+				if (i != this.maxBreakTime)
 				{
-					this.entity.increaseStackSize(stack);
+					this.world.destroyBlockInWorldPartially(this.entity.entityId, breakX, breakY, breakZ, i);
 				}
 
-				this.world.setBlockToAir(breakX, breakY, breakZ);
-				this.world.playAuxSFX(1012, breakX, breakY, breakZ, 0);
+				if (this.breakingTime == 240)
+				{
+					List<ItemStack> droppedStacks = Block.blocksList[blockID].getBlockDropped(world, breakX, breakY, breakZ, this.world.getBlockMetadata(breakX, breakY, breakZ), 0);
 
-				this.resetTask();
+					for (ItemStack stack : droppedStacks)
+					{
+						this.entity.increaseStackSize(stack);
+					}
+
+					this.world.setBlockToAir(breakX, breakY, breakZ);
+					this.world.playAuxSFX(1012, breakX, breakY, breakZ, 0);
+
+					this.resetTask();
+				}
 			}
 		}
-		// this.entity.getNavigator().tryMoveToXYZ(this.moveSpeed);
 	}
 }
